Style: Divide ThumbnailDiskCache
- ThumbnailDiskCache 클래스를 별도의 파일로 분리
